Finished homework

+++ controllers/contacts.js
@@ -6,7 +6,8 @@ const httpError = require("../helpers/httpError");
 
 const listContacts = async (req, res, next) => {
   try {
-    const result = await Contact.find();
+    const { _id: owner } = req.user;
+    const result = await Contact.find({ owner });
     res.json(result);
   } catch (error) {
     next(error);
@@ -15,10 +16,14 @@ const listContacts = async (req, res, next) => {
 
 const getContactById = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
     const result = await Contact.findById(contactId);
     if (!result) {
       throw httpError(404, "Not found!");
+    }
+    if (result.owner.toString() !== owner.toString()) {
+      throw httpError(404, "Not found!!!");
     }
     res.json(result);
   } catch (error) {
@@ -28,12 +33,13 @@ const getContactById = async (req, res, next) => {
 
 const addContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const verifiedResult = schemas.addSchema.validate(req.body);
     if (verifiedResult.error !== undefined) {
       throw httpError(400, "missing required name field");
     }
     const reqBody = verifiedResult.value;
-    const result = await Contact.create(reqBody);
+    const result = await Contact.create({ ...reqBody, owner });
     res.status(201).json(result);
   } catch (error) {
     next(error);
@@ -42,11 +48,13 @@ const addContact = async (req, res, next) => {
 
 const removeContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
-    const result = await Contact.findByIdAndDelete(contactId);
-    if (!result) {
-      throw httpError(404, "Not found!");
+    const contact = await Contact.findById(contactId);
+    if (!contact || contact.owner.toString() !== owner.toString()) {
+      throw httpError(404, "Not found!!!");
     }
+    await Contact.findByIdAndDelete(contactId);
     res.json({ message: "contact deleted" });
   } catch (error) {
     next(error);
@@ -55,16 +63,20 @@ const removeContact = async (req, res, next) => {
 
 const updateContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
+    const contact = await Contact.findById(contactId);
+    if (!contact || contact.owner.toString() !== owner.toString()) {
+      throw httpError(404, "Not found!!!");
+    }
     const verifiedResult = schemas.addSchema.validate(req.body);
     if (verifiedResult.error !== undefined) {
       throw httpError(400, "missing fields");
     }
     const reqBody = verifiedResult.value;
-    const result = await Contact.findByIdAndUpdate(contactId, reqBody, {new:true});
-    if (!result) {
-      throw httpError(404, "Not found!");
-    }
+    const result = await Contact.findByIdAndUpdate(contactId, reqBody, {
+      new: true,
+    });
     res.json(result);
   } catch (error) {
     next(error);
@@ -73,7 +85,12 @@ const updateContact = async (req, res, next) => {
 
 const updateStatusContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
+    const contact = await Contact.findById(contactId);
+    if (!contact || contact.owner.toString() !== owner.toString()) {
+      throw httpError(404, "Not found!!!");
+    }
     const verifiedResult = schemas.updateFavoriteSchema.validate(req.body);
     if (verifiedResult.error !== undefined) {
       throw httpError(400, "missing field favorite");
@@ -82,9 +99,6 @@ const updateStatusContact = async (req, res, next) => {
     const result = await Contact.findByIdAndUpdate(contactId, reqBody, {
       new: true,
     });
-    if (!result) {
-      throw httpError(404, "Not found!");
-    }
     res.json(result);
   } catch (error) {
     next(error);
